Updating matgen library with new code changes and requested code changes in PR comments.
