+++ matgen/generate_matrix_he_and_tz.cc
@@ -1,11 +1,10 @@
-// Copyright (c) 2017-2022, University of Tennessee. All rights reserved.
+// Copyright (c) 2017-2023, University of Tennessee. All rights reserved.
 // SPDX-License-Identifier: BSD-3-Clause
 // This program is free software: you can redistribute it and/or modify it under
 // the terms of the BSD 3-Clause license. See the accompanying LICENSE file.
 
 #include "slate/slate.hh"
-#include "test.hh"
-#include "print_matrix.hh"
+#include "../test/test.hh"
 
 #include <exception>
 #include <string>
@@ -19,44 +18,14 @@
 #include <cstdlib>
 #include <utility>
 
-#include "matrix_params.hh"
+#include "../test/matrix_params.hh"
 #include "generate_matrix.hh"
-#include "random.hh"
+#include "../test/random.hh"
+#include "generate_matrix_utils.hh"
+#include "generate_sigma.hh"
+#include "set_lambdas.hh"
 
 namespace slate {
-
-// -----------------------------------------------------------------------------
-/// Generates an m-by-n Hermitian-storage test matrix.
-/// Handles Hermitian matrices.
-/// Diagonal elements of a Hermitian matrix must be real;
-/// their imaginary part must be 0.
-/// @see generate_matrix
-/// @ingroup generate_matrix
-///
-template <typename scalar_t>
-void generate_matrix(
-    MatrixParams& params,
-    slate::HermitianMatrix<scalar_t>& A,
-    std::vector< blas::real_type<scalar_t> >& Sigma,
-    slate::Options const& opts)
-{
-    slate::BaseTrapezoidMatrix<scalar_t>& TZ = A;
-    generate_matrix( params, TZ, Sigma, opts );
-
-    // Set diagonal to real.
-    #pragma omp parallel for
-    for (int64_t i = 0; i < A.mt(); ++i) {
-        if (A.tileIsLocal( i, i )) {
-            A.tileGetForWriting( i, i, LayoutConvert::ColMajor );
-            auto T = A( i, i );
-            int64_t mb = T.mb();
-            for (int64_t ii = 0; ii < mb; ++ii) {
-                T.at( ii, ii ) = std::real( T( ii, ii ) );
-            }
-        }
-    }
-    A.tileUpdateAllOrigin();
-}
 
 // -----------------------------------------------------------------------------
 /// Generates an m-by-n trapezoid-storage test matrix.
@@ -71,6 +40,7 @@ void generate_matrix(
     std::vector< blas::real_type<scalar_t> >& Sigma,
     slate::Options const& opts)
 {
+    using entry_type = std::function< scalar_t (int64_t, int64_t) >;
     using real_t = blas::real_type<scalar_t>;
 
     // Constants
@@ -324,10 +294,44 @@ void generate_matrix(
     A.tileUpdateAllOrigin();
 }
 
+// -----------------------------------------------------------------------------
+/// Generates an m-by-n Hermitian-storage test matrix.
+/// Handles Hermitian matrices.
+/// Diagonal elements of a Hermitian matrix must be real;
+/// their imaginary part must be 0.
+/// @see generate_matrix
+/// @ingroup generate_matrix
+///
+template <typename scalar_t>
+void generate_matrix(
+    MatrixParams& params,
+    slate::HermitianMatrix<scalar_t>& A,
+    std::vector< blas::real_type<scalar_t> >& Sigma,
+    slate::Options const& opts)
+{
+    slate::BaseTrapezoidMatrix<scalar_t>& TZ = A;
+    generate_matrix( params, TZ, Sigma, opts );
+
+    // Set diagonal to real.
+    #pragma omp parallel for
+    for (int64_t i = 0; i < A.mt(); ++i) {
+        if (A.tileIsLocal( i, i )) {
+            A.tileGetForWriting( i, i, LayoutConvert::ColMajor );
+            auto T = A( i, i );
+            int64_t mb = T.mb();
+            for (int64_t ii = 0; ii < mb; ++ii) {
+                T.at( ii, ii ) = std::real( T( ii, ii ) );
+            }
+        }
+    }
+    A.tileUpdateAllOrigin();
+}
+
+// -----------------------------------------------------------------------------
 /// Overload without Sigma.
 /// @see generate_matrix()
 /// @ingroup generate_matrix
-///----------------------------------------
+///
 template <typename scalar_t>
 void generate_matrix(
     MatrixParams& params,
@@ -350,7 +354,8 @@ void generate_matrix(
     generate_matrix( params, A, dummy, opts );
 }
 
-//----------------------------------------
+//------------------------------------------------------------------------------
+// Explicit instantiations - hermitian matrix.
 template
 void generate_matrix(
     MatrixParams& params,
@@ -375,7 +380,8 @@ void generate_matrix(
     slate::HermitianMatrix< std::complex<double> >& A,
     slate::Options const& opts);
 
-//----------------------------------------
+//------------------------------------------------------------------------------
+// Explicit insantiations - trapezoid matrix.
 template
 void generate_matrix(
     MatrixParams& params,
