+++ matgen/generate_type_jordan.hh
@@ -1,4 +1,4 @@
-// Copyright (c) 2017-2022, University of Tennessee. All rights reserved.
+// Copyright (c) 2017-2023, University of Tennessee. All rights reserved.
 // SPDX-License-Identifier: BSD-3-Clause
 // This program is free software: you can redistribute it and/or modify it under
 // the terms of the BSD 3-Clause license. See the accompanying LICENSE file.
@@ -8,8 +8,7 @@
 
 
 #include "slate/slate.hh"
-#include "test.hh"
-#include "print_matrix.hh"
+#include "../test/test.hh"
 
 #include <exception>
 #include <string>
@@ -23,9 +22,9 @@
 #include <cstdlib>
 #include <utility>
 
-#include "matrix_params.hh"
+#include "../test/matrix_params.hh"
 #include "generate_matrix.hh"
-#include "random.hh"
+#include "../test/random.hh"
 
 
 namespace slate {
