--- test/matrix_generator.cc
@@ -1,2193 +1,0 @@
-// Copyright (c) 2017-2023, University of Tennessee. All rights reserved.
-// SPDX-License-Identifier: BSD-3-Clause
-// This program is free software: you can redistribute it and/or modify it under
-// the terms of the BSD 3-Clause license. See the accompanying LICENSE file.
-
-#include "slate/slate.hh"
-#include "test.hh"
-#include "print_matrix.hh"
-
-#include <exception>
-#include <string>
-#include <vector>
-#include <limits>
-#include <complex>
-#include <chrono>
-
-#include <cmath>
-#include <cstdio>
-#include <cstdlib>
-#include <utility>
-
-#include "matrix_params.hh"
-#include "matrix_generator.hh"
-#include "random.hh"
-
-// -----------------------------------------------------------------------------
-
-enum class TestMatrixType {
-    rand  = int(slate::random::Dist::Uniform),
-    rands = int(slate::random::Dist::UniformSigned),
-    randn = int(slate::random::Dist::Normal),
-    randb = int(slate::random::Dist::Binary),
-    randr = int(slate::random::Dist::BinarySigned),
-    zeros,
-    ones,
-    identity,
-    ij,
-    jordan,
-    chebspec,
-    circul,
-    fiedler,
-    gfpp,
-    kms,
-    orthog,
-    riemann,
-    ris,
-    zielkeNS,
-    diag,
-    svd,
-    poev,
-    heev,
-    geev,
-    geevx,
-};
-
-enum class TestMatrixDist {
-    rand  = int(slate::random::Dist::Uniform),
-    rands = int(slate::random::Dist::UniformSigned),
-    randn = int(slate::random::Dist::Normal),
-    arith,
-    geo,
-    cluster0,
-    cluster1,
-    rarith,
-    rgeo,
-    rcluster0,
-    rcluster1,
-    logrand,
-    specified,
-    none,
-};
-
-const double inf = std::numeric_limits<double>::infinity();
-
-// -----------------------------------------------------------------------------
-// ANSI color codes
-using testsweeper::ansi_esc;
-using testsweeper::ansi_red;
-using testsweeper::ansi_bold;
-using testsweeper::ansi_normal;
-
-// -----------------------------------------------------------------------------
-/// Splits a string by any of the delimiters.
-/// Adjacent delimiters will give empty tokens.
-/// See https://stackoverflow.com/questions/53849
-/// @ingroup util
-std::vector< std::string >
-    split( const std::string& str, const std::string& delims )
-{
-    size_t npos = std::string::npos;
-    std::vector< std::string > tokens;
-    size_t start = (str.size() > 0 ? 0 : npos);
-    while (start != npos) {
-        size_t end = str.find_first_of( delims, start );
-        tokens.push_back( str.substr( start, end - start ));
-        start = (end == npos ? npos : end + 1);
-    }
-    return tokens;
-}
-
-// =============================================================================
-namespace slate {
-
-// -----------------------------------------------------------------------------
-/// Generates Sigma vector of singular or eigenvalues, according to distribution.
-///
-/// Internal function, called from generate_matrix().
-///
-/// @ingroup generate_matrix
-///
-template <typename matrix_type>
-void generate_sigma(
-    MatrixParams& params,
-    TestMatrixDist dist, bool rand_sign,
-    blas::real_type<typename matrix_type::value_type> cond,
-    blas::real_type<typename matrix_type::value_type> sigma_max,
-    matrix_type& A,
-    std::vector< blas::real_type<typename matrix_type::value_type> >& Sigma,
-    int64_t seed )
-{
-    using scalar_t = typename matrix_type::value_type;
-    using real_t = blas::real_type<scalar_t>;
-
-    // Constants
-    const scalar_t zero = 0.0;
-
-    // locals
-    int64_t min_mn = std::min( A.m(), A.n() );
-
-    // Ensure Sigma is allocated
-    if (Sigma.size() == 0) {
-        Sigma.resize(min_mn);
-    }
-    assert( min_mn == int64_t(Sigma.size()) );
-
-    switch (dist) {
-        case TestMatrixDist::arith:
-            for (int64_t i = 0; i < min_mn; ++i) {
-                Sigma[i] = 1 - i / real_t(min_mn - 1) * (1 - 1/cond);
-            }
-            break;
-
-        case TestMatrixDist::rarith:
-            for (int64_t i = 0; i < min_mn; ++i) {
-                Sigma[i] = 1 - (min_mn - 1 - i) / real_t(min_mn - 1) * (1 - 1/cond);
-            }
-            break;
-
-        case TestMatrixDist::geo:
-            for (int64_t i = 0; i < min_mn; ++i) {
-                Sigma[i] = pow( cond, -i / real_t(min_mn - 1) );
-            }
-            break;
-
-        case TestMatrixDist::rgeo:
-            for (int64_t i = 0; i < min_mn; ++i) {
-                Sigma[i] = pow( cond, -(min_mn - 1 - i) / real_t(min_mn - 1) );
-            }
-            break;
-
-        case TestMatrixDist::cluster0:
-            Sigma[0] = 1;
-            for (int64_t i = 1; i < min_mn; ++i) {
-                Sigma[i] = 1/cond;
-            }
-            break;
-
-        case TestMatrixDist::rcluster0:
-            for (int64_t i = 0; i < min_mn-1; ++i) {
-                Sigma[i] = 1/cond;
-            }
-            Sigma[min_mn-1] = 1;
-            break;
-
-        case TestMatrixDist::cluster1:
-            for (int64_t i = 0; i < min_mn-1; ++i) {
-                Sigma[i] = 1;
-            }
-            Sigma[min_mn-1] = 1/cond;
-            break;
-
-        case TestMatrixDist::rcluster1:
-            Sigma[0] = 1/cond;
-            for (int64_t i = 1; i < min_mn; ++i) {
-                Sigma[i] = 1;
-            }
-            break;
-
-        case TestMatrixDist::logrand: {
-            real_t range = log( 1/cond );
-            slate::random::generate(slate::random::Dist::Uniform, seed,
-                                    Sigma.size(), 1, 0, 0,
-                                    Sigma.data(), Sigma.size());
-            for (int64_t i = 0; i < min_mn; ++i) {
-                Sigma[i] = exp( Sigma[i] * range );
-            }
-            // make cond exact
-            if (min_mn >= 2) {
-                Sigma[0] = 1;
-                Sigma[1] = 1/cond;
-            }
-            break;
-        }
-
-        case TestMatrixDist::randn:
-        case TestMatrixDist::rands:
-        case TestMatrixDist::rand: {
-            auto rand_dist = slate::random::Dist(int(dist));
-            slate::random::generate(rand_dist, seed,
-                                    Sigma.size(), 1, 0, 0,
-                                    Sigma.data(), 1);
-            break;
-        }
-
-        case TestMatrixDist::specified:
-            // user-specified Sigma values; don't modify
-            sigma_max = 1;
-            rand_sign = false;
-            break;
-
-        case TestMatrixDist::none:
-            assert( false );
-            break;
-    }
-
-    if (sigma_max != 1) {
-        blas::scal( Sigma.size(), sigma_max, Sigma.data(), 1 );
-    }
-
-    if (rand_sign) {
-        // apply random signs
-        for (int64_t i = 0; i < min_mn; ++i) {
-            float rand;
-            // use j_offset==1 to get different random values than above
-            slate::random::generate(slate::random::Dist::Uniform, seed,
-                                    1, 1, i, 1,
-                                    &rand, 1);
-            if (rand > 0.5) {
-                Sigma[i] = -Sigma[i];
-            }
-        }
-    }
-
-    // copy Sigma => A
-    int64_t min_mt_nt = std::min(A.mt(), A.nt());
-    set(zero, zero, A);
-    int64_t S_index = 0;
-    #pragma omp parallel
-    #pragma omp master
-    for (int64_t i = 0; i < min_mt_nt; ++i) {
-        if (A.tileIsLocal(i, i)) {
-            #pragma omp task slate_omp_default_none shared( A, Sigma ) \
-                firstprivate( i, S_index )
-            {
-                A.tileGetForWriting( i, i, LayoutConvert::ColMajor );
-                auto T = A(i, i);
-                for (int ii = 0; ii < A.tileNb(i); ++ii) {
-                    T.at(ii, ii) = Sigma[S_index + ii];
-                }
-            }
-        }
-        S_index += A.tileNb(i);
-    }
-
-    A.tileUpdateAllOrigin();
-}
-
-
-// -----------------------------------------------------------------------------
-/// Given matrix A with singular values such that sum(sigma_i^2) = n,
-/// returns A with columns of unit norm, with the same condition number.
-/// see: Davies and Higham, 2000, Numerically stable generation of correlation
-/// matrices and their factors.
-///
-/// Internal function, called from generate_matrix().
-///
-/// @ingroup generate_matrix
-//template <typename scalar_t>
-//void generate_correlation_factor( slate::Matrix<scalar_t>& A )
-//{
-//    //const scalar_t eps = std::numeric_limits<scalar_t>::epsilon();
-//
-//    std::vector<scalar_t> x( A.n );
-//    for (int64_t j = 0; j < A.n; ++j) {
-//        x[j] = blas::dot( A.m, A(0,j), 1, A(0,j), 1 );
-//    }
-//
-//    for (int64_t i = 0; i < A.n; ++i) {
-//        for (int64_t j = 0; j < A.n; ++j) {
-//            if ((x[i] < 1 && 1 < x[j]) || (x[i] > 1 && 1 > x[j])) {
-//                scalar_t xij, d, t, c, s;
-//                xij = blas::dot( A.m, A(0,i), 1, A(0,j), 1 );
-//                d = sqrt( xij*xij - (x[i] - 1)*(x[j] - 1) );
-//                t = (xij + std::copysign( d, xij )) / (x[j] - 1);
-//                c = 1 / sqrt(1 + t*t);
-//                s = c*t;
-//                blas::rot( A.m, A(0,i), 1, A(0,j), 1, c, -s );
-//                x[i] = blas::dot( A.m, A(0,i), 1, A(0,i), 1 );
-//                //if (x[i] - 1 > 30*eps) {
-//                //    printf( "i %d, x[i] %.6f, x[i] - 1 %.6e, 30*eps %.6e\n",
-//                //            i, x[i], x[i] - 1, 30*eps );
-//                //}
-//                //assert( x[i] - 1 < 30*eps );
-//                x[i] = 1;
-//                x[j] = blas::dot( A.m, A(0,j), 1, A(0,j), 1 );
-//                break;
-//            }
-//        }
-//    }
-//}
-//
-//
-// -----------------------------------------------------------------------------
-// specialization to complex
-// can't use Higham's algorithm in complex
-//template<>
-//void generate_correlation_factor( slate::Matrix<std::complex<float>>& A )
-//{
-//    assert( false );
-//}
-//
-//template<>
-//void generate_correlation_factor( slate::Matrix<std::complex<double>>& A )
-//{
-//    assert( false );
-//}
-
-
-// -----------------------------------------------------------------------------
-/// Generates matrix using SVD, $A = U Sigma V^H$.
-///
-/// Internal function, called from generate_matrix().
-///
-/// @ingroup generate_matrix
-template <typename scalar_t>
-void generate_svd(
-    MatrixParams& params,
-    TestMatrixDist dist,
-    blas::real_type<scalar_t> cond,
-    blas::real_type<scalar_t> condD,
-    blas::real_type<scalar_t> sigma_max,
-    slate::Matrix<scalar_t>& A,
-    std::vector< blas::real_type<scalar_t> >& Sigma,
-    int64_t seed,
-    slate::Options const& opts )
-{
-    using real_t = blas::real_type<scalar_t>;
-    assert( A.m() >= A.n() );
-
-    // locals
-    int64_t n = A.n();
-    int64_t mt = A.mt();
-    int64_t nt = A.nt();
-    int64_t min_mt_nt = std::min(mt, nt);
-
-    slate::Matrix<scalar_t> U = A.emptyLike();
-    U.insertLocalTiles();
-    slate::TriangularFactors<scalar_t> T;
-
-    // ----------
-    generate_sigma( params, dist, false, cond, sigma_max, A, Sigma, seed );
-    seed += 1;
-
-    // for generate correlation factor, need sum sigma_i^2 = n
-    // scaling doesn't change cond
-    if (condD != 1) {
-        real_t sum_sq = blas::dot( Sigma.size(), Sigma.data(), 1, Sigma.data(), 1 );
-        real_t scale = sqrt( Sigma.size() / sum_sq );
-        blas::scal( Sigma.size(), scale, Sigma.data(), 1 );
-
-        // copy Sigma to diag(A)
-        int64_t S_index = 0;
-        #pragma omp parallel
-        #pragma omp master
-        for (int64_t i = 0; i < min_mt_nt; ++i) {
-            if (A.tileIsLocal(i, i)) {
-                #pragma omp task slate_omp_default_none shared( A, Sigma ) \
-                    firstprivate( i, S_index )
-                {
-                    A.tileGetForWriting( i, i, LayoutConvert::ColMajor );
-                    auto Aii = A(i, i);
-                    for (int ii = 0; ii < A.tileNb(i); ++ii) {
-                        Aii.at(ii, ii) = Sigma[S_index + ii];
-                    }
-                }
-            }
-            S_index += A.tileNb(i);
-        }
-    }
-
-    // random U, m-by-min_mn
-    #pragma omp parallel
-    #pragma omp master
-    {
-        int64_t j_global = 0;
-        for (int64_t j = 0; j < nt; ++j) {
-            int64_t i_global = 0;
-            for (int64_t i = 0; i < mt; ++i) {
-                if (A.tileIsLocal(i, j)) {
-                    #pragma omp task slate_omp_default_none shared( U ) \
-                        firstprivate( i, j, i_global, j_global, seed )
-                    {
-                        U.tileGetForWriting( i, j, LayoutConvert::ColMajor );
-                        auto Uij = U(i, j);
-                        slate::random::generate(slate::random::Dist::Normal, seed,
-                                                Uij.mb(), Uij.nb(), i_global, j_global,
-                                                Uij.data(), Uij.stride());
-                    }
-                }
-                i_global += A.tileMb(i);
-            }
-            j_global += A.tileNb(j);
-        }
-    }
-    seed += 1;
-
-    // we need to make each random column into a Householder vector;
-    // no need to update subsequent columns (as in geqrf).
-    // However, currently we do geqrf here,
-    // since we don't have a way to make Householder vectors (no distributed larfg).
-    slate::geqrf(U, T, opts);
-
-    // A = U*A
-    slate::unmqr( slate::Side::Left, slate::Op::NoTrans, U, T, A, opts);
-
-    // random V, n-by-min_mn (stored column-wise in U)
-    auto V = U.slice(0, n-1, 0, n-1);
-    int64_t V_mt = V.mt();
-    int64_t V_nt = V.nt();
-
-    #pragma omp parallel
-    #pragma omp master
-    {
-        int64_t j_global = 0;
-        for (int64_t j = 0; j < V_nt; ++j) {
-            int64_t i_global = 0;
-            for (int64_t i = 0; i < V_mt; ++i) {
-                if (A.tileIsLocal(i, j)) {
-                    #pragma omp task slate_omp_default_none shared( V ) \
-                        firstprivate( i, j, i_global, j_global, seed )
-                    {
-                        V.tileGetForWriting( i, j, LayoutConvert::ColMajor );
-                        auto Vij = V(i, j);
-                        slate::random::generate(slate::random::Dist::Normal, seed,
-                                                Vij.mb(), Vij.nb(), i_global, j_global,
-                                                Vij.data(), Vij.stride());
-                    }
-                }
-                i_global += A.tileMb(i);
-            }
-            j_global += A.tileNb(j);
-        }
-    }
-    seed += 1;
-
-    slate::geqrf(V, T, opts);
-
-    // A = A*V^H
-    slate::unmqr( slate::Side::Right, slate::Op::ConjTrans, V, T, A, opts);
-
-    if (condD != 1) {
-        // A = A*W, W orthogonal, such that A has unit column norms,
-        // i.e., A^H A is a correlation matrix with unit diagonal
-        // TODO: uncomment generate_correlation_factor
-        //generate_correlation_factor( A );
-
-        // A = A*D col scaling
-        std::vector<real_t> D( n );
-        real_t range = log( condD );
-        slate::random::generate(slate::random::Dist::Uniform, seed,
-                                n, 1, 0, 0,
-                                D.data(), n);
-        for (auto& D_i: D) {
-            D_i = exp( D_i * range );
-        }
-        // TODO: add argument to return D to caller?
-        if (params.verbose) {
-            printf( "D = [" );
-            for (auto& D_i: D) {
-                printf( " %11.8g", D_i );
-            }
-            printf( " ];\n" );
-        }
-
-        int64_t J_index = 0;
-        for (int64_t j = 0; j < nt; ++j) {
-            for (int64_t i = 0; i < mt; ++i) {
-                if (A.tileIsLocal(i, j)) {
-                    A.tileGetForWriting( i, j, LayoutConvert::ColMajor );
-                    auto Aij = A(i, j);
-                    for (int jj = 0; jj < A.tileNb(j); ++jj) {
-                        for (int ii = 0; ii < A.tileMb(i); ++ii) {
-                            Aij.at(ii, jj) *= D[J_index + jj];
-                        }
-                    }
-                }
-            }
-            J_index += A.tileNb(j);
-        }
-    }
-    A.tileUpdateAllOrigin();
-}
-
-// -----------------------------------------------------------------------------
-/// Generates matrix using Hermitian eigenvalue decomposition, $A = U Sigma U^H$.
-///
-/// Internal function, called from generate_matrix().
-///
-/// @ingroup generate_matrix
-template <typename scalar_t>
-void generate_heev(
-    MatrixParams& params,
-    TestMatrixDist dist, bool rand_sign,
-    blas::real_type<scalar_t> cond,
-    blas::real_type<scalar_t> condD,
-    blas::real_type<scalar_t> sigma_max,
-    slate::Matrix<scalar_t>& A,
-    std::vector< blas::real_type<scalar_t> >& Sigma,
-    int64_t seed,
-    slate::Options const& opts )
-{
-    using real_t = blas::real_type<scalar_t>;
-
-    // check inputs
-    assert( A.m() == A.n() );
-
-    // locals
-    int64_t n = A.n();
-    slate::Matrix<scalar_t> U = A.emptyLike();
-    U.insertLocalTiles();
-    slate::TriangularFactors<scalar_t> T;
-
-    // ----------
-    generate_sigma( params, dist, rand_sign, cond, sigma_max, A, Sigma, seed );
-    seed += 1;
-
-    // random U, m-by-min_mn
-    int64_t nt = U.nt();
-    int64_t mt = U.mt();
-
-    #pragma omp parallel
-    #pragma omp master
-    {
-        int64_t j_global = 0;
-        for (int64_t j = 0; j < nt; ++j) {
-            int64_t i_global = 0;
-            for (int64_t i = 0; i < mt; ++i) {
-                if (A.tileIsLocal(i, j)) {
-                    #pragma omp task slate_omp_default_none shared( U ) \
-                        firstprivate( i, j, i_global, j_global, seed )
-                    {
-                        U.tileGetForWriting( i, j, LayoutConvert::ColMajor );
-                        auto Uij = U(i, j);
-                        slate::random::generate(slate::random::Dist::Normal, seed,
-                                                Uij.mb(), Uij.nb(), i_global, j_global,
-                                                Uij.data(), Uij.stride());
-                    }
-                }
-                i_global += A.tileMb(i);
-            }
-            j_global += A.tileNb(j);
-        }
-    }
-    seed += 1;
-
-    // we need to make each random column into a Householder vector;
-    // no need to update subsequent columns (as in geqrf).
-    // However, currently we do geqrf here,
-    // since we don't have a way to make Householder vectors (no distributed larfg).
-    slate::geqrf(U, T, opts);
-
-    // A = U*A
-    slate::unmqr( slate::Side::Left, slate::Op::NoTrans, U, T, A, opts );
-
-    // A = A*U^H
-    slate::unmqr( slate::Side::Right, slate::Op::ConjTrans, U, T, A, opts );
-
-    // make diagonal real
-    // usually LAPACK ignores imaginary part anyway, but Matlab doesn't
-    #pragma omp parallel for
-    for (int64_t i = 0; i < nt; ++i) {
-        if (A.tileIsLocal(i, i)) {
-            A.tileGetForWriting( i, i, LayoutConvert::ColMajor );
-            auto Aii = A(i, i);
-            for (int ii = 0; ii < A.tileMb(i); ++ii) {
-                Aii.at(ii, ii) = std::real( Aii.at(ii, ii) );
-            }
-        }
-    }
-
-    if (condD != 1) {
-        // A = D*A*D row & column scaling
-        std::vector<real_t> D( n );
-        real_t range = log( condD );
-        slate::random::generate(slate::random::Dist::Uniform, seed,
-                                n, 1, 0, 0,
-                                D.data(), n);
-        for (int64_t i = 0; i < n; ++i) {
-            D[i] = exp( D[i] * range );
-        }
-
-        int64_t J_index = 0;
-        #pragma omp parallel
-        #pragma omp master
-        for (int64_t j = 0; j < nt; ++j) {
-            int64_t I_index = 0;
-            for (int64_t i = 0; i < mt; ++i) {
-                if (A.tileIsLocal(i, j)) {
-                    #pragma omp task slate_omp_default_none shared( A, D ) \
-                        firstprivate( i, j, I_index, J_index )
-                    {
-                        A.tileGetForWriting( i, j, LayoutConvert::ColMajor );
-                        auto Aij = A(i, j);
-                        for (int jj = 0; jj < A.tileMb(j); ++jj) {
-                            for (int ii = 0; ii < A.tileMb(i); ++ii) {
-                                Aij.at(ii, jj) *= D[I_index + ii] * D[J_index + jj];
-                            }
-                        }
-                    }
-                }
-                I_index += A.tileNb(i);
-            }
-            J_index += A.tileMb(j);
-        }
-    }
-    A.tileUpdateAllOrigin();
-}
-
-// -----------------------------------------------------------------------------
-/// Generates matrix using general eigenvalue decomposition, $A = V T V^H$,
-/// with orthogonal eigenvectors.
-/// Not yet implemented.
-///
-/// Internal function, called from generate_matrix().
-///
-/// @ingroup generate_matrix
-template <typename scalar_t>
-void generate_geev(
-    MatrixParams& params,
-    TestMatrixDist dist,
-    blas::real_type<scalar_t> cond,
-    blas::real_type<scalar_t> sigma_max,
-    slate::Matrix<scalar_t>& A,
-    std::vector< blas::real_type<scalar_t> >& Sigma,
-    int64_t seed,
-    slate::Options const& opts )
-{
-    throw std::exception();  // not implemented
-}
-
-// -----------------------------------------------------------------------------
-/// Generates matrix using general eigenvalue decomposition, $A = X T X^{-1}$,
-/// with random eigenvectors.
-/// Not yet implemented.
-///
-/// Internal function, called from generate_matrix().
-///
-/// @ingroup generate_matrix
-template <typename scalar_t>
-void generate_geevx(
-    MatrixParams& params,
-    TestMatrixDist dist,
-    blas::real_type<scalar_t> cond,
-    blas::real_type<scalar_t> sigma_max,
-    slate::Matrix<scalar_t>& A,
-    std::vector< blas::real_type<scalar_t> >& Sigma,
-    int64_t seed,
-    slate::Options const& opts )
-{
-    throw std::exception();  // not implemented
-}
-
-// -----------------------------------------------------------------------------
-void generate_matrix_usage()
-{
-    printf(
-    "The --matrix, --cond, and --condD parameters specify a test matrix.\n"
-    "See Test routines: generate_matrix in the HTML documentation for a\n"
-    "complete description.\n"
-    "\n"
-    "%s--matrix%s is one of following:\n"
-    "\n"
-    "%sMatrix%s    |  %sDescription%s\n"
-    "----------|-------------\n"
-    "zeros     |  all zero\n"
-    "ones      |  all one\n"
-    "identity  |  ones on diagonal, rest zero\n"
-    "ij        |  Aij = i + j / 10^ceil( log10( max( m, n ) ) )\n"
-    "jordan    |  ones on diagonal and first subdiagonal, rest zero\n"
-    "chebspec  |  non-singular Chebyshev spectral differentiation matrix\n"
-    "circul    |  circulant matrix where the first column is [1, 2, ..., n]^T\n"
-    "fiedler   |  matrix entry i,j equal to |i - j|\n"
-    "gfpp      |  growth factor for gesv of 1.5^n\n"
-    "kms       |  Kac-Murdock-Szego Toeplitz matrix\n"
-    "orthog    |  matrix entry i,j equal to sqrt(2/(n+1))sin((i+1)(j+1)pi/(n+1))\n"
-    "riemann   |  matrix entry i,j equal to i+1 if j+2 divides i+2 else -1\n"
-    "ris       |  matrix entry i,j equal to 0.5/(n-i-j+1.5)\n"
-    "zielkeNS  |  nonsymmetric matrix of Zielke\n"
-    "          |  \n"
-    "rand@     |  matrix entries random uniform on (0, 1)\n"
-    "rands@    |  matrix entries random uniform on (-1, 1)\n"
-    "randn@    |  matrix entries random normal with mean 0, std 1\n"
-    "randb@    |  matrix entries random uniform from {0, 1}\n"
-    "randr@    |  matrix entries random uniform from {-1, 1}\n"
-    "          |  \n"
-    "diag^@    |  A = Sigma\n"
-    "svd^@     |  A = U Sigma V^H\n"
-    "poev^@    |  A = V Sigma V^H  (eigenvalues positive, i.e., matrix SPD)\n"
-    "spd^@     |  alias for poev\n"
-    "heev^@    |  A = V Lambda V^H (eigenvalues mixed signs)\n"
-    "syev^@    |  alias for heev\n"
-    "geev^@    |  A = V T V^H, Schur-form T                       [not yet implemented]\n"
-    "geevx^@   |  A = X T X^{-1}, Schur-form T, X ill-conditioned [not yet implemented]\n"
-    "\n"
-    "^ and @ denote optional suffixes described below.\n"
-    "\n"
-    "%s^ Distribution%s  |  %sDescription%s\n"
-    "----------------|-------------\n"
-    "_logrand        |  log(sigma_i) random uniform on [ log(1/cond), log(1) ]; default\n"
-    "_arith          |  sigma_i = 1 - frac{i - 1}{n - 1} (1 - 1/cond); arithmetic: sigma_{i+1} - sigma_i is constant\n"
-    "_geo            |  sigma_i = (cond)^{ -(i-1)/(n-1) };             geometric:  sigma_{i+1} / sigma_i is constant\n"
-    "_cluster0       |  Sigma = [ 1, 1/cond, ..., 1/cond ];  1  unit value,  n-1 small values\n"
-    "_cluster1       |  Sigma = [ 1, ..., 1, 1/cond ];      n-1 unit values,  1  small value\n"
-    "_rarith         |  _arith,    reversed order\n"
-    "_rgeo           |  _geo,      reversed order\n"
-    "_rcluster0      |  _cluster0,  reversed order\n"
-    "_rcluster1      |  _cluster1, reversed order\n"
-    "_specified      |  user specified Sigma on input\n"
-    "                |  \n"
-    "_rand           |  sigma_i random uniform on (0, 1)\n"
-    "_rands          |  sigma_i random uniform on (-1, 1)\n"
-    "_randn          |  sigma_i random normal with mean 0, std 1\n"
-    "\n"
-    "%s@ Scaling%s       |  %sDescription%s\n"
-    "----------------|-------------\n"
-    "_ufl            |  scale near underflow         = 1e-308 for double\n"
-    "_ofl            |  scale near overflow          = 2e+308 for double\n"
-    "_small          |  scale near sqrt( underflow ) = 1e-154 for double\n"
-    "_large          |  scale near sqrt( overflow  ) = 6e+153 for double\n"
-    "\n"
-    "%s@ Modifier%s      |  %sDescription%s\n"
-    "----------------|-------------\n"
-    "_dominant       |  make matrix diagonally dominant\n"
-    "_zerocolN       |  set column N to zero, 0 <= N < n\n"
-    "_zerocolFRAC    |  set column N = FRAC * (n-1) to zero, 0 <= FRAC <= 1.0\n"
-    "                |  For Hermitian and symmetric matrices (and currently any\n"
-    "                |  trapezoid matrix), sets row and column N to zero.\n"
-    "\n",
-        ansi_bold, ansi_normal,
-        ansi_bold, ansi_normal,
-        ansi_bold, ansi_normal,
-        ansi_bold, ansi_normal,
-        ansi_bold, ansi_normal,
-        ansi_bold, ansi_normal,
-        ansi_bold, ansi_normal,
-        ansi_bold, ansi_normal,
-        ansi_bold, ansi_normal
-    );
-}
-
-// -----------------------------------------------------------------------------
-/// Decode matrix type, distribution, scaling and modifier.
-///
-template <typename scalar_t>
-void decode_matrix(
-    MatrixParams& params,
-    BaseMatrix<scalar_t>& A,
-    TestMatrixType& type,
-    TestMatrixDist& dist,
-    blas::real_type<scalar_t>& cond,
-    blas::real_type<scalar_t>& condD,
-    blas::real_type<scalar_t>& sigma_max,
-    bool& dominant,
-    int64_t& zero_col )
-{
-    using real_t = blas::real_type<scalar_t>;
-
-    const real_t ufl = std::numeric_limits< real_t >::min();      // == lamch("safe min")  ==  1e-38 or  2e-308
-    const real_t ofl = 1 / ufl;                                   //                            8e37 or   4e307
-    const real_t eps = std::numeric_limits< real_t >::epsilon();  // == lamch("precision") == 1.2e-7 or 2.2e-16
-
-    // locals
-    char msg[ 256 ];
-    std::string kind = params.kind();
-
-    //---------------
-    cond = params.cond_request();
-    bool cond_default = std::isnan( cond );
-    if (cond_default) {
-        cond = 1 / sqrt( eps );
-    }
-
-    condD = params.condD();
-    bool condD_default = std::isnan( condD );
-    if (condD_default) {
-        condD = 1;
-    }
-
-    //---------------
-    sigma_max = 1;
-    std::vector< std::string > tokens = split( kind, "-_" );
-
-    // ----- decode matrix type
-    auto token_iter = tokens.begin();
-    if (token_iter == tokens.end()) {
-        throw std::runtime_error( "empty matrix kind" );
-    }
-    std::string base = *token_iter;
-    ++token_iter;
-    type = TestMatrixType::identity;
-    if      (base == "zeros"   ) { type = TestMatrixType::zeros;    }
-    else if (base == "ones"    ) { type = TestMatrixType::ones;     }
-    else if (base == "identity") { type = TestMatrixType::identity; }
-    else if (base == "ij"      ) { type = TestMatrixType::ij;       }
-    else if (base == "jordan"  ) { type = TestMatrixType::jordan;   }
-    else if (base == "chebspec") { type = TestMatrixType::chebspec; }
-    else if (base == "circul"  ) { type = TestMatrixType::circul;   }
-    else if (base == "fiedler" ) { type = TestMatrixType::fiedler;  }
-    else if (base == "gfpp"    ) { type = TestMatrixType::gfpp;     }
-    else if (base == "kms"     ) { type = TestMatrixType::kms;      }
-    else if (base == "orthog"  ) { type = TestMatrixType::orthog;   }
-    else if (base == "riemann" ) { type = TestMatrixType::riemann;  }
-    else if (base == "ris"     ) { type = TestMatrixType::ris;      }
-    else if (base == "zielkeNS") { type = TestMatrixType::zielkeNS; }
-    else if (base == "randb"   ) { type = TestMatrixType::randb;    }
-    else if (base == "randr"   ) { type = TestMatrixType::randr;    }
-    else if (base == "randn"   ) { type = TestMatrixType::randn;    }
-    else if (base == "rands"   ) { type = TestMatrixType::rands;    }
-    else if (base == "rand"    ) { type = TestMatrixType::rand;     }
-    else if (base == "diag"    ) { type = TestMatrixType::diag;     }
-    else if (base == "svd"     ) { type = TestMatrixType::svd;      }
-    else if (base == "poev" ||
-             base == "spd"     ) { type = TestMatrixType::poev;     }
-    else if (base == "heev" ||
-             base == "syev"    ) { type = TestMatrixType::heev;     }
-    else if (base == "geevx"   ) { type = TestMatrixType::geevx;    }
-    else if (base == "geev"    ) { type = TestMatrixType::geev;     }
-    else {
-        snprintf( msg, sizeof( msg ), "in '%s': unknown matrix '%s'",
-                  kind.c_str(), base.c_str() );
-        throw std::runtime_error( msg );
-    }
-
-    std::string token;
-    dist      = TestMatrixDist::none;
-    sigma_max = 1;
-    dominant  = false;
-    zero_col  = -1;
-
-    while (token_iter != tokens.end()) {
-        token = *token_iter;
-
-        // ----- decode distribution
-        if      (token == "randn"    ) { dist = TestMatrixDist::randn;     }
-        else if (token == "rands"    ) { dist = TestMatrixDist::rands;     }
-        else if (token == "rand"     ) { dist = TestMatrixDist::rand;      }
-        else if (token == "logrand"  ) { dist = TestMatrixDist::logrand;   }
-        else if (token == "arith"    ) { dist = TestMatrixDist::arith;     }
-        else if (token == "geo"      ) { dist = TestMatrixDist::geo;       }
-        else if (token == "cluster1" ) { dist = TestMatrixDist::cluster1;  }
-        else if (token == "cluster0" ) { dist = TestMatrixDist::cluster0;  }
-        else if (token == "rarith"   ) { dist = TestMatrixDist::rarith;    }
-        else if (token == "rgeo"     ) { dist = TestMatrixDist::rgeo;      }
-        else if (token == "rcluster1") { dist = TestMatrixDist::rcluster1; }
-        else if (token == "rcluster0") { dist = TestMatrixDist::rcluster0; }
-        else if (token == "specified") { dist = TestMatrixDist::specified; }
-
-        // ----- decode scaling
-        else if (token == "small") { sigma_max = sqrt( ufl ); }
-        else if (token == "large") { sigma_max = sqrt( ofl ); }
-        else if (token == "ufl"  ) { sigma_max = ufl; }
-        else if (token == "ofl"  ) { sigma_max = ofl; }
-
-        // ----- decode modifiers
-        else if (token == "dominant") { dominant = true; }
-        else if (token.find( "zerocol" ) == 0) {
-            // zeroN for integer N, or zeroR for
-            token = token.substr( 7 );  // skip "zerocol"
-            size_t pos;
-            zero_col = std::stoi( token, &pos, 0 );
-            if (pos < token.size()) {
-                double fraction = std::stod( token, &pos );
-                if (pos < token.size()) {
-                    snprintf( msg, sizeof( msg ),
-                              "in '%s': can't parse number after 'zerocol'",
-                              kind.c_str() );
-                    throw std::runtime_error( msg );
-                }
-                if (fraction < 0.0 || fraction > 1.0) {
-                    snprintf( msg, sizeof( msg ),
-                              "in '%s': fraction outside [0.0, 1.0]",
-                              kind.c_str() );
-                    throw std::runtime_error( msg );
-                }
-                zero_col = int64_t( fraction * (A.n() - 1) );
-            }
-            if (zero_col < 0 || zero_col >= A.n()) {
-                snprintf( msg, sizeof(msg),
-                          "in '%s', column index %lld outside [0, n=%lld)",
-                          kind.c_str(), llong( zero_col ), llong( A.n() ) );
-                throw std::runtime_error( msg );
-            }
-        }
-        else {
-            snprintf( msg, sizeof( msg ), "in '%s': unknown suffix '%s'",
-                      kind.c_str(), token.c_str() );
-            throw std::runtime_error( msg );
-        }
-
-        ++token_iter;
-    }
-
-    // Validate distribution.
-    if (dist != TestMatrixDist::none) {
-        // Error if matrix type doesn't support distribution.
-        if (! (type == TestMatrixType::diag
-               || type == TestMatrixType::svd
-               || type == TestMatrixType::poev
-               || type == TestMatrixType::heev
-               || type == TestMatrixType::geev
-               || type == TestMatrixType::geevx)) {
-            snprintf( msg, sizeof( msg ),
-                      "in '%s': matrix '%s' doesn't support distribution",
-                      kind.c_str(), base.c_str() );
-            throw std::runtime_error( msg );
-        }
-    }
-    else {
-        dist = TestMatrixDist::logrand;  // default
-    }
-
-    // Error if matrix type doesn't support scaling.
-    if (sigma_max != 1
-        && ! (type == TestMatrixType::rand
-              || type == TestMatrixType::rands
-              || type == TestMatrixType::randn
-              || type == TestMatrixType::randb
-              || type == TestMatrixType::randr
-              || type == TestMatrixType::svd
-              || type == TestMatrixType::poev
-              || type == TestMatrixType::heev
-              || type == TestMatrixType::geev
-              || type == TestMatrixType::geevx)) {
-        snprintf( msg, sizeof( msg ),
-                  "in '%s': matrix '%s' doesn't support scaling",
-                  kind.c_str(), base.c_str() );
-        throw std::runtime_error( msg );
-    }
-
-    // Error if matrix type doesn't support diagonally dominant.
-    if (dominant
-        && ! (type == TestMatrixType::rand
-              || type == TestMatrixType::rands
-              || type == TestMatrixType::randn
-              || type == TestMatrixType::randb
-              || type == TestMatrixType::randr
-              || type == TestMatrixType::svd
-              || type == TestMatrixType::poev
-              || type == TestMatrixType::heev
-              || type == TestMatrixType::geev
-              || type == TestMatrixType::geevx))
-    {
-        snprintf( msg, sizeof( msg ),
-                  "in '%s': matrix '%s' doesn't support diagonally dominant",
-                  kind.c_str(), base.c_str() );
-        throw std::runtime_error( msg );
-    }
-
-    // ----- check compatability of options
-    if (A.m() != A.n() &&
-        (type == TestMatrixType::jordan ||
-         type == TestMatrixType::poev   ||
-         type == TestMatrixType::heev   ||
-         type == TestMatrixType::geev   ||
-         type == TestMatrixType::geevx))
-    {
-        snprintf( msg, sizeof( msg ), "in '%s': matrix '%s' requires m == n",
-                  kind.c_str(), base.c_str() );
-        throw std::runtime_error( msg );
-    }
-
-    // Check if cond is known or unused.
-    if (type == TestMatrixType::zeros
-        || type == TestMatrixType::ones
-        || zero_col >= 0) {
-        cond = inf;
-    }
-    else if (type == TestMatrixType::identity
-             || type == TestMatrixType::orthog) {
-        cond = 1;
-    }
-    else if (type != TestMatrixType::svd
-             && type != TestMatrixType::heev
-             && type != TestMatrixType::poev
-             && type != TestMatrixType::geev
-             && type != TestMatrixType::geevx) {
-        // cond unused
-        cond = testsweeper::no_data_flag;
-    }
-    params.cond_actual() = cond;
-
-    // Warn if user set condD and matrix type doesn't use it.
-    if (! condD_default
-        && type != TestMatrixType::svd
-        && type != TestMatrixType::heev
-        && type != TestMatrixType::poev)
-    {
-        fprintf( stderr, "%sWarning: matrix '%s' ignores condD %.2e.%s\n",
-                 ansi_red, kind.c_str(), params.condD(), ansi_normal );
-    }
-
-    // Warn if SPD requested, but distribution is not > 0.
-    if (type == TestMatrixType::poev
-        && (dist == TestMatrixDist::rands
-            || dist == TestMatrixDist::randn)) {
-        fprintf( stderr, "%sWarning: matrix '%s' using rands or randn "
-                 "will not generate SPD matrix; use rand instead.%s\n",
-                 ansi_red, kind.c_str(), ansi_normal );
-    }
-
-    if (params.marked())
-        params.generate_label();
-}
-
-// -----------------------------------------------------------------------------
-/// Generates an arbitrary seed that is unlikely to be repeated
-int64_t configure_seed(MPI_Comm comm, int64_t user_seed)
-{
-
-    // if the given seed is -1, generate a new seed
-    if (user_seed == -1) {
-        // use the highest resolution clock as the seed
-        using namespace std::chrono;
-        user_seed = duration_cast<high_resolution_clock::duration>(
-                        high_resolution_clock::now().time_since_epoch()).count();
-    }
-
-    // ensure seeds are uniform across MPI ranks
-    slate_mpi_call(MPI_Bcast(&user_seed, 1, MPI_INT64_T, 0, comm));
-
-    return user_seed;
-}
-
-// -----------------------------------------------------------------------------
-/// Generates an m-by-n test matrix.
-/// Similar to LAPACK's libtmg functionality, but a level 3 BLAS implementation.
-///
-/// @param[in] params
-///     Test matrix parameters. Uses matrix, cond, condD parameters;
-///     see further details.
-///
-/// @param[out] A
-///     Complex array, dimension (lda, n).
-///     On output, the m-by-n test matrix A in an lda-by-n array.
-///
-/// @param[in,out] Sigma
-///     Real array, dimension (min(m,n))
-///     - On input with matrix distribution "_specified",
-///       contains user-specified singular or eigenvalues.
-///     - On output, contains singular or eigenvalues, if known,
-///       else set to NaN. Sigma is not necesarily sorted.
-///
-/// ### Further Details
-///
-/// The **matrix** parameter specifies the matrix kind according to the
-/// tables below. As indicated, kinds take an optional distribution suffix (^)
-/// and an optional scaling and modifier suffix (@).
-/// The default distribution is logrand.
-/// Examples: rand, rand_small, svd_arith, heev_geo_small.
-///
-/// The **cond** parameter specifies the condition number $cond(S)$, where $S$ is either
-/// the singular values $\Sigma$ or the eigenvalues $\Lambda$, as described by the
-/// distributions below. It does not apply to some matrices and distributions.
-/// For geev and geevx, cond(A) is generally much worse than cond(S).
-/// If _dominant is applied, cond(A) generally improves.
-/// By default, cond(S) = sqrt( 1/eps ) = 6.7e7 for double, 2.9e3 for single.
-///
-/// The **condD** parameter specifies the condition number cond(D), where D is
-/// a diagonal scaling matrix [1]. By default, condD = 1. If condD != 1, then:
-/// - For matrix = svd, set $A = A_0 K D$, where $A_0 = U \Sigma V^H$,
-///   $D$ has log-random entries in $[ \log(1/condD), \log(1) ]$, and
-///   $K$ is diagonal such that columns of $B = A_0 K$ have unit norm,
-///   hence $B^T B$ has unit diagonal.
-///
-/// - For matrix = heev, set $A = D A_0 D$, where $A_0 = U \Lambda U^H$,
-///   $D$ has log-random entries in $[ \log(1/condD), \log(1) ]$.
-///   TODO: set $A = D K A_0 K D$ where
-///   $K$ is diagonal such that $B = K A_0 K$ has unit diagonal.
-///
-/// Note using condD changes the singular or eigenvalues of $A$;
-/// on output, Sigma contains the singular or eigenvalues of $A_0$, not of $A$.
-///
-/// Notation used below:
-/// $\Sigma$ is a diagonal matrix with entries $\sigma_i$ for $i = 1, \dots, n$;
-/// $\Lambda$ is a diagonal matrix with entries $\lambda_i = \pm \sigma_i$,
-/// with random sign;
-/// $U$ and $V$ are random orthogonal matrices from the Haar distribution [2],
-/// $X$ is a random matrix.
-///
-/// See LAPACK Working Note (LAWN) 41:\n
-/// Table  5 (Test matrices for the nonsymmetric eigenvalue problem)\n
-/// Table 10 (Test matrices for the symmetric eigenvalue problem)\n
-/// Table 11 (Test matrices for the singular value decomposition)
-///
-/// Matrix   | Description
-/// ---------|------------
-/// zeros    | all zero
-/// ones     | all one
-/// identity | ones on diagonal, rest zero
-/// jordan   | ones on diagonal and first subdiagonal, rest zero
-/// chebspec | Nonsingular Chebyshev spectral differential matrix
-/// circul   | A circulant matrix where the first column is [1, 2, ..., n]^T
-/// fiedler  | A matrix with entry i,j equal to |i - j|
-/// gfpp     | A matrix with a growth factor of 1.5^n for gesv
-/// kms      | Kac-Murdock-Szego Toeplitz matrix
-/// orthog   | A matrix with entry i,j equal to sqrt(2/(n+1))sin((i+1)(j+1)pi/(n+1))
-/// riemann  | A matrix with entry i,j equal to i+1 if j+2 divides i+2 elso -1
-/// ris      | A matrix with entry i,j equal to 0.5/(n-i-j+1.5)
-/// zielkeNS | A nonsymmetric matrix of Zielke
-/// --       | --
-/// rand@    | matrix entries random uniform on (0, 1)
-/// rands@   | matrix entries random uniform on (-1, 1)
-/// randn@   | matrix entries random normal with mean 0, std 1
-/// randb@   | matrix entries random uniform in {0, 1}
-/// randr@   | matrix entries random uniform in {-1, 1}
-/// --       | --
-/// diag^@   | $A = \Sigma$
-/// svd^@    | $A = U \Sigma V^H$
-/// poev^@   | $A = V \Sigma V^H$  (eigenvalues positive, i.e., matrix SPD)
-/// spd^@    | alias for poev
-/// heev^@   | $A = V \Lambda V^H$ (eigenvalues mixed signs)
-/// syev^@   | alias for heev
-/// geev^@   | $A = V T V^H$, Schur-form $T$                         [not yet implemented]
-/// geevx^@  | $A = X T X^{-1}$, Schur-form $T$, $X$ ill-conditioned [not yet implemented]
-///
-/// Note for geev that $cond(\Lambda)$ is specified, where $\Lambda = diag(T)$;
-/// while $cond(T)$ and $cond(A)$ are usually much worse.
-///
-/// ^ and @ denote optional suffixes described below.
-///
-/// ^ Distribution  |   Description
-/// ----------------|--------------
-/// _logrand        |  $\log(\sigma_i)$ random uniform on $[ \log(1/cond), \log(1) ]$; default
-/// _arith          |  $\sigma_i = 1 - \frac{i - 1}{n - 1} (1 - 1/cond)$; arithmetic: $\sigma_{i+1} - \sigma_i$ is constant
-/// _geo            |  $\sigma_i = (cond)^{ -(i-1)/(n-1) }$;              geometric:  $\sigma_{i+1} / \sigma_i$ is constant
-/// _cluster0       |  $\Sigma = [ 1, 1/cond, ..., 1/cond ]$;  1     unit value,  $n-1$ small values
-/// _cluster1       |  $\Sigma = [ 1, ..., 1, 1/cond ]$;       $n-1$ unit values, 1     small value
-/// _rarith         |  _arith,    reversed order
-/// _rgeo           |  _geo,      reversed order
-/// _rcluster0      |  _cluster0,  reversed order
-/// _rcluster1      |  _cluster1, reversed order
-/// _specified      |  user specified Sigma on input
-/// --              |  --
-/// _rand           |  $\sigma_i$ random uniform on (0, 1)
-/// _rands          |  $\sigma_i$ random uniform on (-1, 1)
-/// _randn          |  $\sigma_i$ random normal with mean 0, std 1
-///
-/// Note _rand, _rands, _randn do not use cond; the condition number is random.
-///
-/// Note for _rands and _randn, $\Sigma$ contains negative values.
-/// This means poev_rands and poev_randn will not generate an SPD matrix.
-///
-/// @ Scaling       |  Description
-/// ----------------|-------------
-/// _ufl            |  scale near underflow         = 1e-308 for double
-/// _ofl            |  scale near overflow          = 2e+308 for double
-/// _small          |  scale near sqrt( underflow ) = 1e-154 for double
-/// _large          |  scale near sqrt( overflow  ) = 6e+153 for double
-///
-/// Note scaling changes the singular or eigenvalues, but not the condition number.
-///
-/// @ Modifier      |  Description
-/// ----------------|-------------
-/// _dominant       |  diagonally dominant: set $A_{i,i} = \pm \max_i( \sum_j |A_{i,j}|, \sum_j |A_{j,i}| )$.
-///
-/// Note _dominant changes the singular or eigenvalues, and the condition number.
-///
-/// ### References
-///
-/// [1] Demmel and Veselic, Jacobi's method is more accurate than QR, 1992.
-///
-/// [2] Stewart, The efficient generation of random orthogonal matrices
-///     with an application to condition estimators, 1980.
-///
-/// @ingroup generate_matrix
-template <typename scalar_t>
-void generate_matrix(
-    MatrixParams& params,
-    slate::Matrix<scalar_t>& A,
-    std::vector< blas::real_type<scalar_t> >& Sigma,
-    slate::Options const& opts)
-{
-    using real_t = blas::real_type<scalar_t>;
-
-    // Constants
-    const real_t nan = std::numeric_limits<real_t>::quiet_NaN();
-    const real_t d_zero = 0;
-    const real_t d_one  = 1;
-    const scalar_t zero = 0;
-    const scalar_t one  = 1;
-    const real_t pi = 3.1415926535897932385;
-
-    // ----------
-    // set Sigma to unknown (nan)
-    lapack::laset( lapack::MatrixType::General, Sigma.size(), 1,
-                   nan, nan, Sigma.data(), Sigma.size() );
-
-    char msg[ 256 ];
-    TestMatrixType type;
-    TestMatrixDist dist;
-    real_t cond;
-    real_t condD;
-    real_t sigma_max;
-    bool dominant;
-    int64_t zero_col;
-    decode_matrix<scalar_t>(
-        params, A, type, dist, cond, condD, sigma_max, dominant, zero_col );
-
-    int64_t seed = configure_seed(A.mpiComm(), params.seed());
-
-    int64_t n = A.n();
-    int64_t m = A.m();
-    int64_t nt = A.nt();
-    int64_t mt = A.mt();
-
-    // ----- generate matrix
-    switch (type) {
-        case TestMatrixType::zeros:
-            set(zero, zero, A);
-            lapack::laset( lapack::MatrixType::General, Sigma.size(), 1,
-                d_zero, d_zero, Sigma.data(), Sigma.size() );
-            break;
-
-        case TestMatrixType::ones:
-            set(one, one, A);
-            lapack::laset( lapack::MatrixType::General, Sigma.size(), 1,
-                d_zero, d_one, Sigma.data(), Sigma.size() );
-            if (Sigma.size() >= 1) {
-                Sigma[0] = sqrt(m*n);
-            }
-            break;
-
-        case TestMatrixType::identity:
-            set(zero, one, A);
-            lapack::laset( lapack::MatrixType::General, Sigma.size(), 1,
-                d_one, d_one, Sigma.data(), Sigma.size() );
-            break;
-
-        case TestMatrixType::ij: {
-            // Scale so j*s < 1.
-            real_t s = 1 / pow( 10, ceil( log10( n ) ) );
-            #pragma omp parallel
-            #pragma omp master
-            {
-                int64_t jj = 0;
-                for (int64_t j = 0; j < nt; ++j) {
-                    int64_t ii = 0;
-                    for (int64_t i = 0; i < mt; ++i) {
-                        #pragma omp task slate_omp_default_none shared( A ) \
-                            firstprivate( i, j, ii, jj, s )
-                        {
-                            if (A.tileIsLocal( i, j )) {
-                                A.tileGetForWriting( i, j, LayoutConvert::ColMajor );
-                                auto Aij = A( i, j );
-                                scalar_t* data = Aij.data();
-                                int64_t lda = Aij.stride();
-                                for (int64_t jjj = 0; jjj < Aij.nb(); ++jjj)
-                                    for (int64_t iii = 0; iii < Aij.mb(); ++iii)
-                                        data[ iii + jjj*lda ] = ii + iii + (jj + jjj)*s;
-                            }
-                        }
-                        ii += A.tileMb( i );
-                    }
-                    jj += A.tileNb( j );
-                }
-            }
-            break;
-        }
-
-        case TestMatrixType::jordan: {
-            set(zero, one, A ); // ones on diagonal
-            // ones on sub-diagonal
-            for (int64_t i = 0; i < nt; ++i) {
-                // Set 1 element from sub-diagonal tile to 1.
-                if (i > 0) {
-                    if (A.tileIsLocal(i, i-1)) {
-                        A.tileGetForWriting( i, i-1, LayoutConvert::ColMajor );
-                        auto T = A(i, i-1);
-                        T.at(0, T.nb()-1) = 1.;
-                    }
-                }
-
-                // Set 1 element from sub-diagonal tile to 1.
-                if (A.tileIsLocal(i, i)) {
-                    A.tileGetForWriting( i, i, LayoutConvert::ColMajor );
-                    auto T = A(i, i);
-                    auto len = T.nb();
-                    for (int j = 0; j < len-1; ++j) {
-                        T.at(j+1, j) = 1.;
-                    }
-                }
-            }
-            break;
-        }
-
-        case TestMatrixType::chebspec: {
-            const int64_t max_mn = std::max(m, n);
-            #pragma omp parallel
-            #pragma omp master
-            {
-                int64_t i_global = 0;
-                for (int64_t i = 0; i < mt; ++i) {
-                    const int64_t mb = A.tileMb(i);
-                    int64_t j_global = 0;
-                    for (int64_t j = 0; j < nt; ++j) {
-                        const int64_t nb = A.tileNb(j);
-                        if (A.tileIsLocal(i, j)) {
-                            #pragma omp task slate_omp_default_none shared( A ) \
-                                firstprivate( i, j, mb, nb, i_global, j_global, \
-                                              max_mn, one, pi )
-                            {
-                                A.tileGetForWriting( i, j, LayoutConvert::ColMajor );
-                                auto A_ij = A(i, j);
-                                for (int64_t ii = 0; ii < mb; ++ii) {
-                                    for (int64_t jj = 0; jj < nb; ++jj) {
-
-                                        int64_t iii = i_global+ii;
-                                        int64_t jjj = j_global+jj;
-
-                                        scalar_t x_i = std::cos(pi*(iii+1)/max_mn);
-                                        scalar_t x_j = std::cos(pi*(jjj+1)/max_mn);
-
-                                        scalar_t value;
-                                        if (iii != jjj) {
-                                            scalar_t c_i = i == max_mn-1 ? 2 : 1;
-                                            scalar_t c_j = j == max_mn-1 ? 2 : 1;
-                                            scalar_t sgn = (i+j)%2 == 0 ? 1 : -1; // (-1)^(i+j)
-
-                                            value = sgn*c_i/(c_j * (x_i - x_j));
-                                        }
-                                        else if (iii+1 == max_mn) {
-                                            value = scalar_t(2*max_mn*max_mn + 1) / scalar_t(-6.0);
-                                        }
-                                        else {
-                                            value = scalar_t(-0.5) * x_i / (one-x_i*x_i);
-                                        }
-
-                                        A_ij.at(ii, jj) = value;
-                                    }
-                                }
-                            }
-                        }
-                        j_global += nb;
-                    }
-                    i_global += mb;
-                }
-                #pragma omp taskwait
-            }
-            break;
-        }
-
-        // circulant matrix for the vector 1:n
-        case TestMatrixType::circul: {
-            const int64_t max_mn = std::max(n, m);
-            #pragma omp parallel
-            #pragma omp master
-            {
-                int64_t i_global = 1;
-                for (int64_t i = 0; i < mt; ++i) {
-                    const int64_t mb = A.tileMb(i);
-                    int64_t j_global = 1;
-                    for (int64_t j = 0; j < nt; ++j) {
-                        const int64_t nb = A.tileNb(j);
-                        if (A.tileIsLocal(i, j)) {
-                            #pragma omp task slate_omp_default_none shared( A ) \
-                                firstprivate( i, j, mb, nb, i_global, j_global, max_mn )
-                            {
-                                A.tileGetForWriting( i, j, LayoutConvert::ColMajor );
-                                auto A_ij = A(i, j);
-                                for (int64_t ii = 0; ii < mb; ++ii) {
-                                    for (int64_t jj = 0; jj < nb; ++jj) {
-                                        auto diff = (j_global+jj) - (i_global+ii);
-                                        A_ij.at(ii, jj) = diff
-                                                          + (diff < 0 ? max_mn : 0)
-                                                          + 1;
-                                    }
-                                }
-                            }
-                        }
-                        j_global += nb;
-                    }
-                    i_global += mb;
-                }
-                #pragma omp taskwait
-            }
-            break;
-        }
-
-        case TestMatrixType::fiedler: {
-            #pragma omp parallel
-            #pragma omp master
-            {
-                int64_t i_global = 0;
-                for (int64_t i = 0; i < mt; ++i) {
-                    const int64_t mb = A.tileMb(i);
-                    int64_t j_global = 0;
-                    for (int64_t j = 0; j < nt; ++j) {
-                        const int64_t nb = A.tileNb(j);
-                        if (A.tileIsLocal(i, j)) {
-                            #pragma omp task slate_omp_default_none shared( A ) \
-                                firstprivate( i, j, mb, nb, i_global, j_global )
-                            {
-                                A.tileGetForWriting( i, j, LayoutConvert::ColMajor );
-                                auto A_ij = A(i, j);
-                                for (int64_t ii = 0; ii < mb; ++ii) {
-                                    for (int64_t jj = 0; jj < nb; ++jj) {
-                                        A_ij.at(ii, jj) = std::abs(i_global + ii
-                                                                   - (j_global + jj));
-                                    }
-                                }
-                            }
-                        }
-                        j_global += nb;
-                    }
-                    i_global += mb;
-                }
-                #pragma omp taskwait
-            }
-            break;
-        }
-
-        case TestMatrixType::gfpp: {
-            set(zero, one, A);
-            #pragma omp parallel for collapse(2)
-            for (int64_t i = 0; i < mt; ++i) {
-                for (int64_t j = 0; j < nt; ++j) {
-                    if (A.tileIsLocal(i, j)) {
-                        A.tileGetForWriting( i, j, LayoutConvert::ColMajor );
-                        auto A_ij = A(i, j);
-                        const int64_t mb = A.tileMb(i);
-                        const int64_t nb = A.tileNb(j);
-                        if (i == j) {
-                            A_ij.set(-0.5, one);
-                            for (int64_t ii = 0; ii < mb-1; ++ii) {
-                                for (int64_t jj = ii+1; jj < nb; ++jj) {
-                                    A_ij.at(ii, jj) = -zero;
-                                }
-                            }
-                        }
-                        else if (i < j) {
-                            A_ij.set(zero);
-                        }
-                        else if (i > j) {
-                            A_ij.set(-0.5);
-                        }
-                        if (j == nt-1) {
-                            const int64_t jj = nb - 1;
-                            for (int64_t ii = 0; ii < mb; ++ii) {
-                                A_ij.at(ii, jj) = 1;
-                            }
-                        }
-                    }
-                }
-            }
-            break;
-        }
-
-        case TestMatrixType::kms: {
-            const scalar_t rho = 0.5;
-            #pragma omp parallel
-            #pragma omp master
-            {
-                int64_t i_global = 1;
-                for (int64_t i = 0; i < mt; ++i) {
-                    const int64_t mb = A.tileMb(i);
-                    int64_t j_global = 1;
-                    for (int64_t j = 0; j < nt; ++j) {
-                        const int64_t nb = A.tileNb(j);
-                        if (A.tileIsLocal(i, j)) {
-                            #pragma omp task slate_omp_default_none shared( A ) \
-                                firstprivate( i, j, mb, nb, i_global, j_global, rho )
-                            {
-                                A.tileGetForWriting( i, j, LayoutConvert::ColMajor );
-                                auto A_ij = A(i, j);
-                                for (int64_t ii = 0; ii < mb; ++ii) {
-                                    for (int64_t jj = 0; jj < nb; ++jj) {
-                                        int64_t iii = i_global + ii;
-                                        int64_t jjj = j_global + jj;
-
-                                        A_ij.at(ii, jj) = std::pow(rho, std::abs(iii - jjj));
-                                    }
-                                }
-                            }
-                        }
-                        j_global += nb;
-                    }
-                    i_global += mb;
-                }
-                #pragma omp taskwait
-            }
-            break;
-        }
-
-        case TestMatrixType::orthog: {
-            const int64_t max_mn = std::max(n, m);
-            const scalar_t outer_const = sqrt(scalar_t(2)/scalar_t(max_mn+1));
-            const scalar_t inner_const = pi/scalar_t(max_mn+1);
-            #pragma omp parallel
-            #pragma omp master
-            {
-                int64_t i_global = 1;
-                for (int64_t i = 0; i < mt; ++i) {
-                    const int64_t mb = A.tileMb(i);
-                    int64_t j_global = 1;
-                    for (int64_t j = 0; j < nt; ++j) {
-                        const int64_t nb = A.tileNb(j);
-                        if (A.tileIsLocal(i, j)) {
-                            #pragma omp task slate_omp_default_none shared( A ) \
-                                firstprivate( i, j, mb, nb, i_global, j_global, \
-                                              inner_const, outer_const )
-                            {
-                                A.tileGetForWriting( i, j, LayoutConvert::ColMajor );
-                                auto A_ij = A(i, j);
-                                for (int64_t ii = 0; ii < mb; ++ii) {
-                                    for (int64_t jj = 0; jj < nb; ++jj) {
-                                        scalar_t a = scalar_t(i_global + ii)
-                                                     * scalar_t(j_global + jj)
-                                                     * inner_const;
-                                        A_ij.at(ii, jj) = outer_const*sin(a);
-                                    }
-                                }
-                            }
-                        }
-                        j_global += nb;
-                    }
-                    i_global += mb;
-                }
-                #pragma omp taskwait
-            }
-            break;
-        }
-
-        case TestMatrixType::riemann: {
-            #pragma omp parallel
-            #pragma omp master
-            {
-                int64_t i_global = 0;
-                for (int64_t i = 0; i < mt; ++i) {
-                    const int64_t mb = A.tileMb(i);
-                    int64_t j_global = 0;
-                    for (int64_t j = 0; j < nt; ++j) {
-                        const int64_t nb = A.tileNb(j);
-                        if (A.tileIsLocal(i, j)) {
-                            #pragma omp task slate_omp_default_none shared( A ) \
-                                firstprivate( i, j, mb, nb, i_global, j_global )
-                            {
-                                A.tileGetForWriting( i, j, LayoutConvert::ColMajor );
-                                auto A_ij = A(i, j);
-                                for (int64_t ii = 0; ii < mb; ++ii) {
-                                    for (int64_t jj = 0; jj < nb; ++jj) {
-                                        int64_t B_i = i_global + ii + 2;
-                                        int64_t B_j = j_global + jj + 2;
-                                        if (B_i % B_j == 0) {
-                                            A_ij.at(ii, jj) = B_i - 1;
-                                        }
-                                        else {
-                                            A_ij.at(ii, jj) = -1;
-                                        }
-                                    }
-                                }
-                            }
-                        }
-                        j_global += nb;
-                    }
-                    i_global += mb;
-                }
-                #pragma omp taskwait
-            }
-            break;
-        }
-
-        case TestMatrixType::ris: {
-            const int64_t max_mn = std::max(n, m);
-            #pragma omp parallel
-            #pragma omp master
-            {
-                int64_t i_global = 1;
-                for (int64_t i = 0; i < mt; ++i) {
-                    const int64_t mb = A.tileMb(i);
-                    int64_t j_global = 1;
-                    for (int64_t j = 0; j < nt; ++j) {
-                        const int64_t nb = A.tileNb(j);
-                        if (A.tileIsLocal(i, j)) {
-                            #pragma omp task slate_omp_default_none shared( A ) \
-                                firstprivate( i, j, mb, nb, i_global, j_global, max_mn )
-                            {
-                                A.tileGetForWriting( i, j, LayoutConvert::ColMajor );
-                                auto A_ij = A(i, j);
-                                for (int64_t ii = 0; ii < mb; ++ii) {
-                                    for (int64_t jj = 0; jj < nb; ++jj) {
-                                        A_ij.at(ii, jj) = 0.5 / (max_mn-(i_global+ii)-(j_global+jj)+1.5);
-                                    }
-                                }
-                            }
-                        }
-                        j_global += nb;
-                    }
-                    i_global += mb;
-                }
-                #pragma omp taskwait
-            }
-            break;
-        }
-
-        case TestMatrixType::zielkeNS: {
-            const int64_t max_mn = std::max(n, m);
-            const scalar_t a = 0.0;
-            #pragma omp parallel
-            #pragma omp master
-            {
-                int64_t i_global = 0;
-                for (int64_t i = 0; i < mt; ++i) {
-                    const int64_t mb = A.tileMb(i);
-                    int64_t j_global = 0;
-                    for (int64_t j = 0; j < nt; ++j) {
-                        const int64_t nb = A.tileNb(j);
-                        if (A.tileIsLocal(i, j)) {
-                            #pragma omp task slate_omp_default_none shared( A ) \
-                                firstprivate( i, j, mb, nb, i_global, j_global, \
-                                              max_mn, one, a )
-                            {
-                                A.tileGetForWriting( i, j, LayoutConvert::ColMajor );
-                                auto A_ij = A(i, j);
-                                for (int64_t ii = 0; ii < mb; ++ii) {
-                                    for (int64_t jj = 0; jj < nb; ++jj) {
-                                        int64_t iii = i_global + ii;
-                                        int64_t jjj = j_global + jj;
-
-                                        if (iii < jjj) {
-                                            A_ij.at(ii, jj) = a+one;
-                                        }
-                                        else if (iii+1 == max_mn && jjj == 0) {
-                                            A_ij.at(ii, jj) = a-one;
-                                        }
-                                        else {
-                                            A_ij.at(ii, jj) = a;
-                                        }
-                                    }
-                                }
-                            }
-                        }
-                        j_global += nb;
-                    }
-                    i_global += mb;
-                }
-                #pragma omp taskwait
-            }
-            break;
-        }
-
-        case TestMatrixType::rand:
-        case TestMatrixType::rands:
-        case TestMatrixType::randn:
-        case TestMatrixType::randb:
-        case TestMatrixType::randr: {
-            auto rand_dist = slate::random::Dist(int(type));
-            #pragma omp parallel
-            #pragma omp master
-            {
-                int64_t j_global = 0;
-                for (int64_t j = 0; j < nt; ++j) {
-                    int64_t i_global = 0;
-                    for (int64_t i = 0; i < mt; ++i) {
-                        if (A.tileIsLocal(i, j)) {
-                            #pragma omp task slate_omp_default_none shared( A ) \
-                                firstprivate( i, j, j_global, i_global, dominant, \
-                                              n, sigma_max, seed, rand_dist )
-                            {
-                                A.tileGetForWriting( i, j, LayoutConvert::ColMajor );
-                                auto Aij = A(i, j);
-                                slate::random::generate(rand_dist, seed,
-                                                        Aij.mb(), Aij.nb(), i_global, j_global,
-                                                        Aij.data(), Aij.stride());
-
-                                // Make it diagonally dominant
-                                if (dominant) {
-                                    if (i == j) {
-                                        int bound = std::min( Aij.mb(), Aij.nb() );
-                                        for (int ii = 0; ii < bound; ++ii) {
-                                            Aij.at(ii, ii) += n;
-                                        }
-                                    }
-                                }
-                                // Scale the matrix
-                                if (sigma_max != 1) {
-                                    scalar_t s = sigma_max;
-                                    tile::scale( s, Aij );
-                                }
-                            }
-                        }
-                        i_global += A.tileMb(i);
-                    }
-                    j_global += A.tileNb(j);
-                }
-            }
-            break;
-        }
-
-        case TestMatrixType::diag:
-            generate_sigma( params, dist, false, cond, sigma_max, A, Sigma, seed );
-            break;
-
-        case TestMatrixType::svd:
-            generate_svd( params, dist, cond, condD, sigma_max, A, Sigma, seed, opts );
-            break;
-
-        case TestMatrixType::poev:
-            generate_heev( params, dist, false, cond, condD, sigma_max, A, Sigma, seed, opts );
-            break;
-
-        case TestMatrixType::heev:
-            generate_heev( params, dist, true, cond, condD, sigma_max, A, Sigma, seed, opts );
-            break;
-
-        case TestMatrixType::geev:
-            generate_geev( params, dist, cond, sigma_max, A, Sigma, seed, opts );
-            break;
-
-        case TestMatrixType::geevx:
-            generate_geevx( params, dist, cond, sigma_max, A, Sigma, seed, opts );
-            break;
-    }
-
-    if (! (type == TestMatrixType::rand  ||
-           type == TestMatrixType::rands ||
-           type == TestMatrixType::randn ||
-           type == TestMatrixType::randb) && dominant) {
-        // make diagonally dominant; strict unless diagonal has zeros
-        snprintf( msg, sizeof( msg ), "in '%s': dominant not yet implemented",
-                  params.kind().c_str() );
-        throw std::runtime_error( msg );
-    }
-
-    // Set col A( :, zero_col ) = 0.
-    if (zero_col >= 0) {
-        #pragma omp parallel
-        #pragma omp master
-        {
-            int64_t j_global = 0;
-            for (int64_t j = 0; j < nt; ++j) {
-                int64_t jj = zero_col - j_global;
-                if (0 <= jj && jj < A.tileNb( j )) {
-                    for (int64_t i = 0; i < mt; ++i) {
-                        if (A.tileIsLocal( i, j )) {
-                            #pragma omp task slate_omp_default_none shared( A ) \
-                                firstprivate( i, j, jj, zero )
-                            {
-                                A.tileGetForWriting( i, j, LayoutConvert::ColMajor );
-                                auto Aij = A( i, j );
-                                lapack::laset(
-                                    lapack::MatrixType::General,
-                                    Aij.mb(), 1, zero, zero,
-                                    &Aij.at( 0, jj ), Aij.stride() );
-                            }
-                        }
-                    }
-                }
-                j_global += A.tileNb( j );
-            }
-        }
-    }
-
-    A.tileUpdateAllOrigin();
-}
-
-// -----------------------------------------------------------------------------
-/// Generates an m-by-n trapezoid-storage test matrix.
-/// Handles Trapezoid, Triangular, Symmetric, and Hermitian matrices.
-/// @see generate_matrix
-/// @ingroup generate_matrix
-///
-template <typename scalar_t>
-void generate_matrix(
-    MatrixParams& params,
-    slate::BaseTrapezoidMatrix<scalar_t>& A,
-    std::vector< blas::real_type<scalar_t> >& Sigma,
-    slate::Options const& opts)
-{
-    using real_t = blas::real_type<scalar_t>;
-
-    // Constants
-    const real_t nan = std::numeric_limits<real_t>::quiet_NaN();
-    const real_t d_zero = 0;
-    const real_t d_one  = 1;
-    const scalar_t zero = 0;
-    const scalar_t one  = 1;
-
-    // ----------
-    // set Sigma to unknown (nan)
-    lapack::laset( lapack::MatrixType::General, Sigma.size(), 1,
-                   nan, nan, Sigma.data(), Sigma.size() );
-
-    char msg[ 256 ];
-    TestMatrixType type;
-    TestMatrixDist dist;
-    real_t cond;
-    real_t condD;
-    real_t sigma_max;
-    bool dominant;
-    int64_t zero_col;
-    decode_matrix<scalar_t>(
-        params, A, type, dist, cond, condD, sigma_max, dominant, zero_col );
-
-    int64_t seed = configure_seed(A.mpiComm(), params.seed());
-
-    int64_t n = A.n();
-    int64_t nt = A.nt();
-    int64_t mt = A.mt();
-
-    // ----- generate matrix
-    switch (type) {
-        case TestMatrixType::zeros:
-            set(zero, zero, A);
-            lapack::laset( lapack::MatrixType::General, Sigma.size(), 1,
-                d_zero, d_zero, Sigma.data(), Sigma.size() );
-            break;
-
-        case TestMatrixType::ones:
-            set(one, one, A);
-            lapack::laset( lapack::MatrixType::General, Sigma.size(), 1,
-                d_zero, d_one, Sigma.data(), Sigma.size() );
-            if (Sigma.size() >= 1) {
-                Sigma[0] = n;
-            }
-            break;
-
-        case TestMatrixType::identity:
-            set(zero, one, A);
-            lapack::laset( lapack::MatrixType::General, Sigma.size(), 1,
-                d_one, d_one, Sigma.data(), Sigma.size() );
-            break;
-
-        case TestMatrixType::jordan: {
-            set(zero, one, A ); // ones on diagonal
-            if (A.uplo() == Uplo::Lower) {
-                // ones on sub-diagonal
-                for (int64_t i = 0; i < nt; ++i) {
-                    // Set 1 element from sub-diagonal tile to 1.
-                    if (i > 0) {
-                        if (A.tileIsLocal(i, i-1)) {
-                            A.tileGetForWriting( i, i-1, LayoutConvert::ColMajor );
-                            auto T = A(i, i-1);
-                            T.at(0, T.nb()-1) = 1.;
-                        }
-                    }
-
-                    // Set 1 element from sub-diagonal tile to 1.
-                    if (A.tileIsLocal(i, i)) {
-                        A.tileGetForWriting( i, i, LayoutConvert::ColMajor );
-                        auto T = A(i, i);
-                        auto len = T.nb();
-                        for (int j = 0; j < len-1; ++j) {
-                            T.at(j+1, j) = 1.;
-                        }
-                    }
-                }
-            }
-            else { // upper
-                // ones on sub-diagonal
-                for (int64_t i = 0; i < nt; ++i) {
-                    // Set 1 element from sub-diagonal tile to 1.
-                    if (i > 0) {
-                        if (A.tileIsLocal(i-1, i)) {
-                            A.tileGetForWriting( i-1, i, LayoutConvert::ColMajor );
-                            auto T = A(i-1, i);
-                            T.at(T.nb()-1, 0) = 1.;
-                        }
-                    }
-
-                    // Set 1 element from sub-diagonal tile to 1.
-                    if (A.tileIsLocal(i, i)) {
-                        A.tileGetForWriting( i, i, LayoutConvert::ColMajor );
-                        auto T = A(i, i);
-                        auto len = T.nb();
-                        for (int j = 0; j < len-1; ++j) {
-                            T.at(j, j+1) = 1.;
-                        }
-                    }
-                }
-            }
-            break;
-        }
-
-        case TestMatrixType::rand:
-        case TestMatrixType::rands:
-        case TestMatrixType::randn:
-        case TestMatrixType::randb:
-        case TestMatrixType::randr: {
-            auto rand_dist = slate::random::Dist(int(type));
-            #pragma omp parallel
-            #pragma omp master
-            {
-                int64_t j_global = 0;
-                for (int64_t j = 0; j < nt; ++j) {
-                    int64_t i_global = 0;
-                    // lower trapezoid is [ j .. mt )
-                    // upper trapezoid is [ 0 .. min( j+1, mt ) )
-                    int64_t i_start = A.uplo() == Uplo::Lower ? j  : 0;
-                    int64_t i_end   = A.uplo() == Uplo::Lower ? mt : std::min( j+1, mt );
-                    for (int64_t i = i_start; i < i_end; ++i) {
-                        if (A.tileIsLocal( i, j )) {
-                            #pragma omp task slate_omp_default_none shared( A ) \
-                                firstprivate( i, j, j_global, i_global, dominant, \
-                                              n, sigma_max, seed, rand_dist )
-                            {
-                                A.tileGetForWriting( i, j, LayoutConvert::ColMajor );
-                                auto Aij = A( i, j );
-                                slate::random::generate(
-                                    rand_dist, seed,
-                                    Aij.mb(), Aij.nb(), i_global, j_global,
-                                    Aij.data(), Aij.stride() );
-
-                                // Make it diagonally dominant
-                                if (dominant && i == j) {
-                                    int bound = std::min( Aij.mb(), Aij.nb() );
-                                    for (int ii = 0; ii < bound; ++ii) {
-                                        Aij.at( ii, ii ) += n;
-                                    }
-                                }
-                                // Scale the matrix
-                                if (sigma_max != 1) {
-                                    scalar_t s = sigma_max;
-                                    tile::scale( s, Aij );
-                                }
-                            }
-                        }
-                        i_global += A.tileMb( i );
-                    }
-                    j_global += A.tileNb( j );
-                }
-            }
-            break;
-        }
-
-        case TestMatrixType::diag:
-            generate_sigma( params, dist, false, cond, sigma_max, A, Sigma, seed );
-            break;
-
-        case TestMatrixType::poev:
-        case TestMatrixType::heev:
-        default:
-            snprintf( msg, sizeof( msg ), "'%s' not yet implemented",
-                      params.kind().c_str() );
-            throw std::runtime_error( msg );
-    }
-
-    if (! (type == TestMatrixType::rand  ||
-           type == TestMatrixType::rands ||
-           type == TestMatrixType::randn ||
-           type == TestMatrixType::randb) && dominant) {
-        // make diagonally dominant; strict unless diagonal has zeros
-        snprintf( msg, sizeof( msg ), "in '%s', dominant not yet implemented",
-                  params.kind().c_str() );
-        throw std::runtime_error( msg );
-    }
-
-    if (zero_col >= 0) {
-        // Set col A( :, zero_col ) = 0.
-        #pragma omp parallel
-        #pragma omp master
-        {
-            int64_t j_global = 0;
-            for (int64_t j = 0; j < nt; ++j) {
-                int64_t jj = zero_col - j_global;
-                if (0 <= jj && jj < A.tileNb( j )) {
-                    // lower trapezoid is i in [ j, mt )
-                    // upper trapezoid is i in [ 0, min( j+1, mt ) )
-                    int64_t i_start = A.uplo() == Uplo::Lower ? j  : 0;
-                    int64_t i_end   = A.uplo() == Uplo::Lower ? mt : std::min( j+1, mt );
-                    for (int64_t i = i_start; i < i_end; ++i) {
-                        if (A.tileIsLocal( i, j )) {
-                            #pragma omp task slate_omp_default_none shared( A ) \
-                                firstprivate( i, j, jj, zero )
-                            {
-                                A.tileGetForWriting( i, j, LayoutConvert::ColMajor );
-                                auto Aij = A( i, j );
-                                lapack::laset(
-                                    lapack::MatrixType::General,
-                                    Aij.mb(), 1, zero, zero,
-                                    &Aij.at( 0, jj ), Aij.stride() );
-                            }
-                        }
-                    }
-                }
-                j_global += A.tileNb( j );
-            }
-        }
-
-        // Set row A( zero_col, : ) = 0.
-        // Merging this with the above omp parallel would require
-        // todo: this should apply only to Hermitian or symmetric matrices,
-        // not to triangular or trapezoid matrices. Perhaps
-        //     generate_matrix( ..., HermitianMatrix, ... )
-        // could set a flag when calling this routine.
-        #pragma omp parallel
-        #pragma omp master
-        {
-            int64_t i_global = 0;
-            for (int64_t i = 0; i < mt; ++i) {
-                int64_t ii = zero_col - i_global;
-                if (0 <= ii && ii < A.tileMb( i )) {
-                    // upper trapezoid is j in [ i, nt )
-                    // lower trapezoid is j in [ 0, min( i+1, nt ) )
-                    int64_t j_start = A.uplo() == Uplo::Upper ? i  : 0;
-                    int64_t j_end   = A.uplo() == Uplo::Upper ? nt : std::min( i+1, nt );
-                    for (int64_t j = j_start; j < j_end; ++j) {
-                        if (A.tileIsLocal( i, j )) {
-                            #pragma omp task slate_omp_default_none shared( A ) \
-                                firstprivate( i, j, ii, zero )
-                            {
-                                A.tileGetForWriting( i, j, LayoutConvert::ColMajor );
-                                auto Aij = A( i, j );
-                                lapack::laset(
-                                    lapack::MatrixType::General,
-                                    1, Aij.nb(), zero, zero,
-                                    &Aij.at( ii, 0 ), Aij.stride() );
-                            }
-                        }
-                    }
-                }
-                i_global += A.tileMb( i );
-            }
-        }
-    }
-
-    A.tileUpdateAllOrigin();
-}
-
-// -----------------------------------------------------------------------------
-/// Generates an m-by-n Hermitian-storage test matrix.
-/// Handles Hermitian matrices.
-/// Diagonal elements of a Hermitian matrix must be real;
-/// their imaginary part must be 0.
-/// @see generate_matrix
-/// @ingroup generate_matrix
-///
-template <typename scalar_t>
-void generate_matrix(
-    MatrixParams& params,
-    slate::HermitianMatrix<scalar_t>& A,
-    std::vector< blas::real_type<scalar_t> >& Sigma,
-    slate::Options const& opts)
-{
-    slate::BaseTrapezoidMatrix<scalar_t>& TZ = A;
-    generate_matrix( params, TZ, Sigma, opts );
-
-    // Set diagonal to real.
-    #pragma omp parallel for
-    for (int64_t i = 0; i < A.mt(); ++i) {
-        if (A.tileIsLocal( i, i )) {
-            A.tileGetForWriting( i, i, LayoutConvert::ColMajor );
-            auto T = A( i, i );
-            int64_t mb = T.mb();
-            for (int64_t ii = 0; ii < mb; ++ii) {
-                T.at( ii, ii ) = std::real( T( ii, ii ) );
-            }
-        }
-    }
-    A.tileUpdateAllOrigin();
-}
-// -----------------------------------------------------------------------------
-/// Overload without Sigma.
-/// @see generate_matrix()
-/// @ingroup generate_matrix
-///
-template <typename scalar_t>
-void generate_matrix(
-    MatrixParams& params,
-    slate::Matrix<scalar_t>& A,
-    slate::Options const& opts)
-{
-    using real_t = blas::real_type<scalar_t>;
-    std::vector<real_t> dummy;
-    generate_matrix( params, A, dummy, opts );
-}
-
-/// Overload without Sigma.
-/// @see generate_matrix()
-/// @ingroup generate_matrix
-///
-template <typename scalar_t>
-void generate_matrix(
-    MatrixParams& params,
-    slate::BaseTrapezoidMatrix<scalar_t>& A,
-    slate::Options const& opts)
-{
-    using real_t = blas::real_type<scalar_t>;
-    std::vector<real_t> dummy;
-    generate_matrix( params, A, dummy, opts );
-}
-
-/// Overload without Sigma.
-/// @see generate_matrix()
-/// @ingroup generate_matrix
-///
-template <typename scalar_t>
-void generate_matrix(
-    MatrixParams& params,
-    slate::HermitianMatrix<scalar_t>& A,
-    slate::Options const& opts)
-{
-    using real_t = blas::real_type<scalar_t>;
-    std::vector<real_t> dummy;
-    generate_matrix( params, A, dummy, opts );
-}
-
-//------------------------------------------------------------------------------
-// Explicit instantiations.
-template
-void generate_matrix(
-    MatrixParams& params,
-    slate::Matrix<float>& A,
-    slate::Options const& opts);
-
-template
-void generate_matrix(
-    MatrixParams& params,
-    slate::Matrix<double>& A,
-    slate::Options const& opts);
-
-template
-void generate_matrix(
-    MatrixParams& params,
-    slate::Matrix< std::complex<float> >& A,
-    slate::Options const& opts);
-
-template
-void generate_matrix(
-    MatrixParams& params,
-    slate::Matrix< std::complex<double> >& A,
-    slate::Options const& opts);
-
-//----------------------------------------
-template
-void generate_matrix(
-    MatrixParams& params,
-    slate::BaseTrapezoidMatrix<float>& A,
-    slate::Options const& opts);
-
-template
-void generate_matrix(
-    MatrixParams& params,
-    slate::BaseTrapezoidMatrix<double>& A,
-    slate::Options const& opts);
-
-template
-void generate_matrix(
-    MatrixParams& params,
-    slate::BaseTrapezoidMatrix< std::complex<float> >& A,
-    slate::Options const& opts);
-
-template
-void generate_matrix(
-    MatrixParams& params,
-    slate::BaseTrapezoidMatrix< std::complex<double> >& A,
-    slate::Options const& opts);
-
-//----------------------------------------
-template
-void generate_matrix(
-    MatrixParams& params,
-    slate::HermitianMatrix<float>& A,
-    slate::Options const& opts);
-
-template
-void generate_matrix(
-    MatrixParams& params,
-    slate::HermitianMatrix<double>& A,
-    slate::Options const& opts);
-
-template
-void generate_matrix(
-    MatrixParams& params,
-    slate::HermitianMatrix< std::complex<float> >& A,
-    slate::Options const& opts);
-
-template
-void generate_matrix(
-    MatrixParams& params,
-    slate::HermitianMatrix< std::complex<double> >& A,
-    slate::Options const& opts);
-
-} // namespace slate
